bare_err_tree
Function print_tree

pub fn print_tree<const FRONT_MAX: usize, E, F>(tree: E, formatter: F) -> Result
where
    E: AsErrTree,
    F: Write,

Produces ErrTree formatted output for an error.

FRONT_MAX limits the number of leading bytes. Each deeper error requires 6 bytes to fit “│ “. So for a max depth of 3 errors, FRONT_MAX == 18. By default, FRONT_MAX bytes are allocated on stack. When heap_buffer is enabled, the bytes are allocated on stack and FRONT_MAX only acts as a depth limit. When tracing is enabled, at most FRONT_MAX stack traces will be tracked for duplicates.

Errors must define Error::source correctly for the tree to display. The derive macros for ErrTree track extra information and handle multiple sources (Error::source is designed around a single error source).

use bare_err_tree::{AsErrTree, print_tree};

const PRINT_SIZE: usize = 60;

fn sized_print<E, F>(tree: E, formatter: F) -> fmt::Result
where
    E: AsErrTree,
    F: fmt::Write,
{
    print_tree::<PRINT_SIZE, _, _>(tree, formatter)
}

let mut out = String::new();
sized_print(&io::Error::last_os_error() as &dyn Error, &mut out).unwrap();
println!("{out}");