bare_err_tree

Trait AsErrTree

pub trait AsErrTree {
    // Required method
    fn as_err_tree(&self, func: &mut dyn FnMut(ErrTree<'_>));
}

Defines an Error’s temporary view as an ErrTree for printing.

This can be defined with err_tree, manually (see ErrTree), or with the default dyn implementation. The dyn implementation does not track any more information than standard library errors or track multiple sources.

Implementors must call func with a properly constructed ErrTree.

Required Methods

fn as_err_tree(&self, func: &mut dyn FnMut(ErrTree<'_>))

Constructs the ErrTree internally and calls func on it.